Merge from Upstream

+++ Makefile
@@ -1,12 +1,32 @@
 
 CWD = $(shell pwd)
+ARCH := $(shell uname -i)
+UNAME_A := $(shell uname -a)
 
 all: module tools java_agent pkg
+ifneq ($(findstring Ubuntu,$(UNAME_A) $(shell test -e /etc/os-release && head -1 /etc/os-release)),)
+	dpkg -P diagnose-tools || echo "remove diagnose-tools error"
+	cd rpmbuild; sudo dpkg -i diagnose-tools*.deb
+else
 	yum remove -y diagnose-tools
 	yum install -y rpmbuild/RPMS/x86_64/diagnose-tools-*.rpm
 	diagnose-tools -v
+endif
 
 devel:
+ifneq ($(findstring Ubuntu,$(UNAME_A) $(shell test -e /etc/os-release && head -1 /etc/os-release)),)
+	apt update
+	apt -y install gcc
+	apt -y install g++
+	apt -y install libunwind8-dev
+	apt -y install elfutils
+	apt -y install libelf-dev
+	apt -y install rpm
+	apt -y install alien
+	apt -y install bash-completion # git自动补全
+	apt -y install --no-install-recommends openjdk-8-jdk
+else
+	yum check-update
 	yum install -y libstdc++-static
 	yum install -y glibc-static
 	yum install -y zlib-devel
@@ -18,6 +38,7 @@ devel:
 	yum install -y rpm-build
 	yum install -y xz-libs
 	yum install -y xz-devel
+endif
 	sh ./vender/devel.sh
 
 deps:
@@ -26,7 +47,6 @@ deps:
 	#cd SOURCE/diagnose-tools/xz; ./autogen.sh; ./configure CFLAGS="-g -O2" --prefix=$(PWD)/SOURCE/diagnose-tools/deps; make install
 	#cd SOURCE/diagnose-tools/zlib; ./configure --prefix=$(PWD)/SOURCE/diagnose-tools/deps; make install
 	cd SOURCE/diagnose-tools/java_agent; make
-
 	sh ./vender/deps.sh
 
 .PHONY: deps
@@ -45,17 +65,20 @@ java_agent:
 pkg:
 	cd rpmbuild; sh rpmbuild.sh
 	ls rpmbuild/RPMS/*/*
-
-deb:
-	rm ./rpmbuild/diagnose-tools*.deb
-	sudo alien -d ./rpmbuild/RPMS/x86_64/diagnose-tools*.rpm
-	#dpkg -P diagnose-tools || echo "remove alibaba diagnose tool error"
-	#sudo dpkg -i diagnose-tools*.deb
+ifneq ($(findstring Ubuntu,$(UNAME_A) $(shell test -e /etc/os-release && head -1 /etc/os-release)),)
+	#sudo dpkg-reconfigure dash !!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+	cd rpmbuild; rm -f diagnose-tools*.deb
+ifneq ($(findstring aarch64,$(ARCH)),)
+	cd rpmbuild; sudo alien -d --target=arm64 ./RPMS/aarch64/diagnose-tools*.rpm
+else
+	cd rpmbuild; sudo alien -d ./RPMS/x86_64/diagnose-tools*.rpm
+endif
+endif
 
 test:
 	modprobe ext4
 	insmod SOURCE/module/diagnose.ko || echo ""
-	sh ./SOURCE/script/test.sh
+	bash ./SOURCE/script/test.sh $(case)
 	rmmod diagnose
 	rm tmp.txt -f
 	rm *.svg -f
